requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AI/ML
langchain==0.1.0
langchain-openai==0.0.5
langgraph==0.0.20
openai==1.5.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
tqdm==4.66.1